requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alabaster==0.7.12
alembic==1.4.1
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.5
astropy==4.0.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
azure-core==1.9.0
azure-storage-blob==12.6.0
Babel==2.8.1
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bitarray==1.6.1
bkcharts==0.2
bleach==3.2.1
blinker==1.4
blis==0.7.3
bokeh==2.2.3
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
catalogue==1.0.0
catboost==0.24.3
certifi==2021.10.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clipboard==0.0.4
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
colorlover==0.3.0
combo==0.1.1
comtypes==1.1.7
conda==4.10.3
conda-build==3.20.5
conda-package-handling==1.7.2
conda-verify==3.4.2
confuse==1.4.0
coverage==5.5
cryptography==3.1.1
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.4
Cython==0.29.24
cytoolz==0.11.0
dask==2.30.0
databricks-cli==0.14.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distributed==2.30.1
docker==4.4.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
flatbuffers==1.12
fsspec==0.8.3
funcy==1.15
future==0.18.2
gast==0.3.3
gensim==3.8.3
geojson==2.5.0
gevent==20.9.0
gitdb==4.0.5
GitPython==3.1.11
glob2==0.7
gmpy2==2.0.8
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.28.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
googletrans==3.0.0
graphviz==0.15
greenlet==0.4.17
grpcio==1.34.1
gTTS==2.2.1
h11==0.9.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
hstspreload==2020.11.21
html5lib==1.1
htmlmin==0.1.12
httpcore==0.9.1
httplib2==0.18.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
imagesize==1.2.0
imbalanced-learn==0.7.0
IMDbPY==2020.9.25
importlib-metadata==2.0.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.3.0
kmodes==0.10.2
lazy-object-proxy==1.4.3
libarchive-c==2.9
lightgbm==3.1.0
livelossplot==0.5.3
llvmlite==0.34.0
locket==0.2.0
lxml==4.6.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
menuinst==1.4.16
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mlflow==1.12.1
mlxtend==0.18.0
mock==4.0.2
more-itertools==8.6.0
MouseInfo==0.1.3
mpmath==1.1.0
msgpack==1.0.0
msrest==0.6.19
multidict==5.1.0
multipledispatch==0.6.0
murmurhash==1.0.4
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.3.56
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.3.3
pandas-profiling==2.9.0
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.2.4
pkginfo==1.6.1
plac==1.1.3
playsound==1.2.2
plotly==4.13.0
pluggy==0.13.1
ply==3.11
preshed==3.0.4
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.8
protobuf==3.17.2
psutil==5.7.2
py==1.9.0
py-translate==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
PyAutoGUI==0.9.52
pycaret==2.2.2
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
PyGetWindow==0.0.9
Pygments==2.7.2
pyjokes==0.6.0
PyJWT==1.7.1
pyLDAvis==2.1.2
pylint==2.6.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyod==0.8.4
pyOpenSSL==19.1.0
pyowm==3.1.1
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
pyqtgraph==0.11.0
pyreadline==2.1
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
PySocks==1.7.1
pytest==0.0.0
python-dateutil==2.8.1
python-editor==1.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
querystring-parser==1.2.4
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
roman==3.3
rope==0.18.0
rsa==4.7.2
Rtree==0.9.4
ruamel-yaml==0.15.87
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.4.1
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
setuptools==50.3.1.post20201107
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.13
six==1.15.0
smart-open==4.0.1
smmap==3.0.4
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
sounddevice==0.4.1
soupsieve==2.0.1
spacy==2.3.4
SpeechRecognition==3.8.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.20
sqlparse==0.4.1
srsly==1.0.4
statsmodels==0.12.0
suod==0.0.4
sympy==1.6.2
tables==3.6.1
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
tblib==1.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.1
testcode==1.0.0
testpath==0.4.4
textblob==0.15.3
thinc==7.4.3
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
toolz==0.11.1
tornado==6.0.4
touch==0.0.0
tqdm==4.50.2
traitlets==5.0.5
twilio==6.59.1
typing-extensions==3.7.4.3
ujson==4.0.1
umap-learn==0.4.6
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
values==0.0.0
visions==0.5.0
waitress==1.4.4
wasabi==0.8.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
win-inet-pton==1.1.0
win-unicode-console==0.5
win10toast==0.9
wincertstore==0.2
winshell==0.6
wolframalpha==4.1.0
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.2.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwings==0.20.8
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
zict==2.0.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
